Added Tons of Grid Functionality
Added get_walkable_cells which will be a key function in unit pathfinding.

It will determine where the unit can move to based on it's movement range and movement type.

It will then highlight the area and make a fancy little border using the UnitOverlay tilemap and some simple overlay_tileset. Uses auto-tile functionality to create the outline.

Probably need to work on the movement logic for all movement types next